+++ backend/canisters/group/impl/src/lib.rs
@@ -109,9 +109,10 @@ impl RuntimeState {
         }
     }
 
-    pub fn push_notification(&mut self, recipients: Vec<UserId>, notification: Notification) {
+    pub fn push_notification(&mut self, sender: Option<UserId>, recipients: Vec<UserId>, notification: Notification) {
         if !recipients.is_empty() {
             let args = c2c_push_notification::Args {
+                sender,
                 recipients,
                 authorizer: Some(self.data.local_group_index_canister_id),
                 notification_bytes: ByteBuf::from(candid::encode_one(notification).unwrap()),